moved update_network from main to sidebar
